+++ packages/types/index.d.ts
@@ -2,3 +2,4 @@ export * from "./src/api";
 export * from "./src/db";
 export * from "./src/moodle";
 export * from "./src/telegram";
+export * from "./src/schedule";
